projects/carwash/app/auth/utils.py
import jwt

from app.core.config import settings

from passlib.context import CryptContext
from datetime import datetime, timedelta, timezone
from typing import Union, Any


pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")


def verify_password(plain_password, hashed_password):
    return pwd_context.verify(plain_password, hashed_password)


def get_password_hash(password):
    return pwd_context.hash(password)


def create_access_token(data: Union[str, dict]):
    if isinstance(data, str):
        data = {"sub": data}  # Обернуть строку в словарь с ключом 'sub'

    to_encode = data.copy()

    expire = datetime.now(timezone.utc) + timedelta(minutes=settings.auth.access_token_expire_minutes)
    to_encode.update({"exp": expire})

    encoded_jwt = jwt.encode(to_encode, settings.auth.secret_key, algorithm=settings.auth.algorithm)
    return encoded_jwt



def create_refresh_token(subject: Union[str, Any], expires_delta: int = None) -> str:
    if expires_delta is not None:
        expires_delta = datetime.now(timezone.utc) + expires_delta
    else:
        expires_delta = datetime.now(timezone.utc) + timedelta(minutes=settings.auth.refresh_token_expire_minutes)

    to_encode = {"exp": expires_delta, "sub": str(subject)}
    encoded_jwt = jwt.encode(to_encode, settings.auth.refresh_secret_key, settings.auth.algorithm)
    return encoded_jwt