projects/carwash/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = {extras = ["email"], version = "^2.7.1"}
pydantic-settings = "^2.2.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
asyncpg = "^0.30.0"
alembic = "^1.13.1"
orjson = "^3.10.11"
aiosqlite = "^0.20.0"
fastapi-filter = "^2.0.1"
fastapi-pagination = {extras = ["sqlalchemy"], version = "^0.12.34"}
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
redis = "^5.2.1"
celery = {extras = ["redis"], version = "^5.4.0"}
email-validator = "^2.2.0"
aiosmtplib = "^3.0.2"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
